refactor(blog): extract CollapsibleTagList to eliminate duplicated tag expand/collapse logic
